fix(v3): engine/data review — streaming import, scalar stages, cursor leaks, argv secrets
Streaming import (whole-file materialisation OOM'd on ~1GB NDJSON; working
set is now one 1000-doc batch, JSON arrays via incremental JsonReader).
Aggregation: scalar-bodied stages ($limit/$count/$unwind/$sortByCount)
parse — 4 of 10 shipped templates previously threw.
Migration: copy cursor closed on pause/cancel/failure; cancel honoured in
index/verify phases; verification counts bounded by maxTime; resume adds
hint+min index bound (MongoDB 4.4 planned $expr as a full index scan —
resume re-read the whole collection); pause/cancel no longer brick start()
via stray map entries; orphan reconcile unconditional (crash + quick
relaunch left jobs stuck 'running').
Backups: mongodump/mongorestore credentials via 0600 --config file instead
of world-readable argv; spawn-failure double-finish removed.
Also: UUID() shell syntax now emits valid EJSON base64, schema sampling no
longer collapses to 1 doc past 2^31 documents, listDatabases size tolerant
of Int/Double, master key created with owner-only perms before writing,
export cursor closed on write failure.

+++ v3/src/main/kotlin/io/mex/mongo/Schema.kt
@@ -34,7 +34,9 @@ private class Stat {
 fun analyzeSchema(client: MongoClient, db: String, coll: String, sampleSize: Int = 1000): SchemaReport {
     val collection = client.getDatabase(db).getCollection(coll)
     val total = collection.estimatedDocumentCount()
-    val sz = minOf(sampleSize, maxOf(total.toInt(), 1))
+    // Compare as Long first: total.toInt() wraps negative past 2^31 docs, which collapsed
+    // the sample to a single document on exactly the collections that matter.
+    val sz = minOf(sampleSize.toLong(), maxOf(total, 1L)).toInt()
     val docs = collection.aggregate(listOf(Document("\$sample", Document("size", sz)))).toList()
     val stats = HashMap<String, Stat>()
     for (doc in docs) walk(doc, "", stats)
